\name{estimateEffects}
\alias{estimateEffects}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Estimate Effects on the Frontier
}
\description{
  \code{estimateEffects()} is used to estimate the effect of the
    treatment along the entire frontier. 
}
\usage{
estimateEffects(frontier.object, formula)
}
\arguments{
  \item{frontier.object}{
    An object generated by \code{makeFrontier()}.
  }
  \item{formula}{
    An object of class formula (or one that can be
    coerced to that class). This will be passed to
    \code{lm()}.
  }
}
\examples{
\dontrun{
data(lalonde)
match.on <- colnames(lalonde)[!(colnames(lalonde) \%in\% c('re78', 'treat'))]
my.frontier <- makeFrontier(dataset = lalonde,
                            treatment = 'treat',
                            outcome = 're78',
                            match.on = match.on)
my.estimates <- estimateEffects(my.frontier, 're78 ~ treat')
}
}