\name{lalonde}
\alias{lalonde}
\docType{data}
\title{Modified Lalonde dataset}
\description{
This is a modified version of the Lalonde experimental dataset used for
explanatory purposes only.
}
\usage{data(lalonde)}
\format{
  A data frame with 1185 observations on the following 11 variables.
  \describe{
    \item{\code{treat}}{treatment variable indicator}
    \item{\code{age}}{age}
    \item{\code{education}}{years of education}
    \item{\code{black}}{race indicator variable}
    \item{\code{married}}{marital status indicator variable}
    \item{\code{nodegree}}{indicator variable for not possessing a degree}
    \item{\code{re74}}{real earnings in 1974}
    \item{\code{re75}}{real earnings in 1975}
    \item{\code{re78}}{real earnings in 1978 (post-treatment outcome)}
    \item{\code{hispanic}}{ethnic indicator variable}
  }
}
